feat: update `Menu` and` MenuItem` components with new styles, props and create `Pokemon` components folder
